SENTRY-1337: Move GroupMappingService from sentry-provider-common to sentry-core-common(Colin Ma, reviewed by Dapeng Sun)
